+++ src/dlt_filesystem/source/api.py
@@ -1,6 +1,7 @@
 from dlt_filesystem.source.impl.local import LocalFilesystemSource
 from dlt_filesystem.source.impl.remote import (
     AzureSource,
+    DropboxSource,
     GCSSource,
     HDFSSource,
     OCISource,
@@ -12,6 +13,7 @@ from dlt_filesystem.source.impl.remote import (
 
 __all__ = [
     "AzureSource",
+    "DropboxSource",
     "GCSSource",
     "HDFSSource",
     "LocalFilesystemSource",